refactor: 피드 헤더 분리
